move Selector widget to its own file and create a class for its Options
I wanted this so I can enable/disable options after creating the widget.
